fix(server): review round 12 — event-clock lineage, restart-proof completion, eject re-check
- webhook envelopes parse LiveKit createdAt; ObservedCallSids carries
  it and ParticipantState records event-clock lineage so a re-executed
  stale join redelivery cannot roll a participant sid backward
- clear_local_state re-checks the live issuance bucket before arming a
  departure eject: a racing initiate that repopulated issued (clearing
  the eject at mint and register) suppresses the arm
- the observed-registration occupied-empty arm refreshes created_at /
  room_sid stamps like the Jingle arm
- CallThreadEndRetry is restart-proof end to end: rows carry the
  persisted end instant (thread_ended_at_ms) reused by every retry;
  they are exempt from the ownership drain gate when the room is
  globally unclaimed; the SFU-unavailable early return honors the
  fence; broadcast is best-effort when no room actor exists anywhere;
  and CallThreadEnd joined the uncapped-retry exemption
- new dedupe index (status, call_id, action) for the enqueue path
- call-thread completion machinery split into routes/call_thread_end.rs
